// src/api/model/worshipDetailResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * 주뜨 청년부 앱 API
 * 주뜨 청년부 앱 REST API 명세
 * OpenAPI spec version: v1
 */
import type { WorshipDetailResponseStatus } from './worshipDetailResponseStatus';
import type { BulletinInfo } from './bulletinInfo';
import type { PraiseInfo } from './praiseInfo';
import type { AnnouncementInfo } from './announcementInfo';

export interface WorshipDetailResponse {
  id?: number;
  sermonTitle?: string;
  worshipAt?: string;
  preacherName?: string;
  verseReference?: string;
  verseText?: string;
  youtubeUrl?: string;
  youtubeVideoId?: string;
  status?: WorshipDetailResponseStatus;
  hasSummary?: boolean;
  bulletins?: BulletinInfo[];
  praises?: PraiseInfo[];
  announcements?: AnnouncementInfo[];
  createdAt?: string;
  updatedAt?: string;
}