// src/api/model/worshipListItemResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * 주뜨 청년부 앱 API
 * 주뜨 청년부 앱 REST API 명세
 * OpenAPI spec version: v1
 */
import type { WorshipListItemResponseStatus } from './worshipListItemResponseStatus';

export interface WorshipListItemResponse {
  id?: number;
  sermonTitle?: string;
  worshipAt?: string;
  preacherName?: string;
  youtubeVideoId?: string;
  status?: WorshipListItemResponseStatus;
}